requirements.txt
asn1crypto==0.24.0
awsebcli==3.15.0
blessed==1.15.0
botocore==1.12.130
cached-property==1.5.1
cachetools==3.1.0
cement==2.8.2
certifi==2018.10.15
cffi==1.12.2
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
decorator==4.3.0
defusedxml==0.5.0
dialogflow==0.5.2
Django==2.1.5
django-allauth==0.38.0
django-coreapi==2.3.0
django-cors-headers==2.4.1
django-filter==2.1.0
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.3
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docker==3.7.2
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
future==0.16.0
google-api-core==1.10.0
google-auth==1.6.3
googleapis-common-protos==1.5.10
grpcio==1.20.1
idna==2.7
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.4
jsonschema==2.6.0
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.1.1
oauthlib==3.0.1
pathspec==0.5.9
plivo==4.1.3
protobuf==3.7.1
psycopg2==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==19.0.0
PySocks==1.6.8
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
PyYAML==5.1
requests==2.20.1
requests-oauthlib==1.2.0
rsa==4.0
semantic-version==2.5.0
six==1.11.0
stripe==2.21.0
termcolor==1.1.0
texttable==0.9.1
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
websocket-client==0.56.0
